Move request types to new file
changelog_begin
changelog_end
